+++ packages/core/di/index.ts
@@ -2,4 +2,3 @@ export * from './injectable';
 export * from './factory-creator';
 export * from './injector';
 export * from './injected';
-export * from './build-private-injector';