+++ packages/core/di/injector.ts
@@ -1,13 +1,34 @@
 /**
- * IOC container for InDiv
- * 
- * methods: push, find, get
+ * IOC container
  *
  * @export
  * @class Injector
  */
 export class Injector {
+  /**
+   * parent injector from forken child
+   *
+   * @private
+   * @type {Injector}
+   * @memberof Injector
+   */
+  public parentInjector: Injector;
+
+  /**
+   * provider map: save provider with key and value
+   *
+   * @private
+   * @type {Map<any, any>}
+   * @memberof Injector
+   */
   private readonly providerMap: Map<any, any> = new Map();
+  /**
+   * instance map: save singleton instance with key and value
+   *
+   * @private
+   * @type {Map<any, any>}
+   * @memberof Injector
+   */
   private readonly instanceMap: Map<any, any> = new Map();
 
   /**
@@ -29,7 +50,9 @@ export class Injector {
    * @memberof Injector
    */
   public getProvider(key: any): any {
-    return this.providerMap.get(key);
+    if (this.providerMap.has(key)) return this.providerMap.get(key);
+    else if (this.parentInjector) return this.parentInjector.getProvider(key);
+    else return undefined;
   }
 
   /**
@@ -40,7 +63,10 @@ export class Injector {
    * @memberof Injector
    */
   public setInstance(key: any, value: any): void {
-    if (!this.instanceMap.has(key)) this.instanceMap.set(key, value);
+    if (this.providerMap.has(key)) {
+      if (!this.instanceMap.has(key)) this.instanceMap.set(key, value);
+    } else if (this.parentInjector) this.parentInjector.setInstance(key, value);
+    else console.error(`injector can'n set instance of provider: ${(key as any).name}`);
   }
 
   /**
@@ -51,8 +77,52 @@ export class Injector {
    * @memberof Injector
    */
   public getInstance(key: any): any {
-    if (this.instanceMap.has(key)) return this.instanceMap.get(key);
-    return null;
+    if (this.providerMap.has(key)) {
+      if (this.instanceMap.has(key)) return this.instanceMap.get(key);
+      else return undefined;
+    } else if (this.parentInjector) return this.parentInjector.getInstance(key);
+    else return undefined;
+  }
+
+  /**
+   * set provider and it's instance in an injector
+   *
+   * @param {*} key
+   * @param {*} provider
+   * @param {*} instance
+   * @memberof Injector
+   */
+  public setProviderAndInstance(key: any, provider: any, instance: any): void {
+    this.setProvider(key, provider);
+    this.setInstance(key, instance);
+  }
+
+  /**
+   * get parent injector of a provider
+   *
+   * @param {*} key
+   * @returns {Injector}
+   * @memberof Injector
+   */
+  public getParentInjectorOfProvider(key: any): Injector {
+    if (this.providerMap.has(key)) return this;
+    else if (this.parentInjector) return this.parentInjector.getParentInjectorOfProvider(key);
+    else {
+      console.error(`injector can't find parent injector of provider: ${(key as any).name}`);
+      return undefined;
+    }
+  }
+
+  /**
+   * fork a child inject from this injector
+   *
+   * @returns {Injector}
+   * @memberof Injector
+   */
+  public fork(): Injector {
+    const childInjector = new Injector();
+    childInjector.parentInjector = this;
+    return childInjector;
   }
 }