+++ packages/core/directive/index.ts
@@ -18,7 +18,6 @@ export function Directive(options: TDirectiveOptions): (_constructor: Function) 
   return function (_constructor: Function): void {
     injected(_constructor);
     (_constructor as any).isSingletonMode = false;
-    rootInjector.setProvider(_constructor, _constructor);
     (_constructor as any).nvType = 'nvDirective';
     (_constructor as any).selector = options.selector;
     const vm: IDirective = _constructor.prototype;